import of CINT 5.15.168. Incorporates Philippe most recent fixes. For what else is new see http://root.cern.ch/root/Cint.phtml?relnote.
git-svn-id: http://root.cern.ch/svn/root/trunk@11320 27541ba8-7e3a-0410-8455-c3a389f83636

+++ cint/src/sizeof.c
@@ -1599,7 +1599,7 @@ struct G__param *libp;
       case 'k':
 	libp->para[i].obj.d = libp->para[i].obj.i;
 	libp->para[i].type = formal_type;
-	libp->para[i].ref = &libp->para[i].obj.d;
+	libp->para[i].ref = (long)(&libp->para[i].obj.d);
 	break;
       }
       break;
@@ -1616,7 +1616,7 @@ struct G__param *libp;
       case 'f':
 	libp->para[i].obj.i = libp->para[i].obj.d;
 	libp->para[i].type = formal_type;
-	libp->para[i].ref = &libp->para[i].obj.i;
+	libp->para[i].ref = (long)(&libp->para[i].obj.i);
 	break;
       }
       break;
